tree: add jenkins-wait-for-build
This script uses the Jenkins remote API to follow the status of a build
until it completes. The job description is printed whenever it changes,
which can be used as a rough way to measure progress. A future
enhancement would be to support streamling logs.

It requires the current kubectl/oc context to be a namespace in which
Jenkins is installed.

I use it in my terminal to get notified when a build I'm waiting for is
done using e.g.:

```
$ jenkins-wait-for-build release 11; bell
```

(`bell` is an alias for `echo -e '\a'`; i.e. emit an alert which will
show up in at least tmux as the window getting lit up in the bottom
status line.)