tree: add jenkins-get-running-builds
This script prints out the list of running builds for a particular job.
It can optionally take parameters to further filter the list. Example
usage:

    $ jenkins-get-running-builds build-arch STREAM=4.10
    #11: [4.10][s390x] ⚡ 410.84.202212051934-0
    #10: [4.10][ppc64le] Running
    #9: [4.10][aarch64] ⚡ 410.84.202212051934-0

    $ jenkins-get-running-builds build-arch STREAM=4.10 ARCH=ppc64le
    #10: [4.10][ppc64le] Running

This has some utility for local devs, but it also demonstrates the APIs
that a higher-level pipeline driver could query to determine whether a
build is already running with a given set of parameters.