tmux.conf: conditionalize prefix and styling on `~/.tmux.conf.nested`
It's useful to have tmux inside a container use the default prefix so
that we don't have to spam the key when we're already running within
tmux. Also keep the default green status bar so it's easier to tell
apart.

This file will be touched by our container setup script.